pyproject.toml
[project]
name = "diffms-yaolab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.3.3",
    "pillow>=11.3.0",
    "pydantic>=2.12.3",
    "rdkit>=2025.9.1",
    "streamlit>=1.50.0",
]
